refactor: 기물 객체의 상속 계층 구조화
